Moved cosmology functions into utils
