--- libs/cosmology/pyproject.toml
@@ -1,20 +1,0 @@
-[tool.poetry]
-name = "cosmology"
-version = "0.1.0"
-description = "Library to ensure cosmology standardization across Aframe"
-authors = ["Will Benoit <benoi090@umn.edu>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-astropy = "^5.0"
-scipy = "^1.7"
-rpds-py = "<0.20"
-
-[tool.poetry.group.dev.dependencies]
-jupyter = "^1.0.0"
-pytest = "^8.2.0"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
