+++ libs/ledger/pyproject.toml
@@ -12,7 +12,7 @@ bilby = "^2.2.2"
 h5py = "^3.10.0"
 numpy = "^1.26"
 
-cosmology = {path = "../cosmology", develop = true}
+utils = {path = "../utils", develop = true}
 
 [tool.poetry.group.dev.dependencies]
 jupyter = "^1.0.0"
