+++ projects/plots/apptainer.def
@@ -7,7 +7,6 @@ Stage: build
 ../../libs/utils /opt/aframe/libs/utils
 ../../libs/ledger /opt/aframe/libs/ledger
 ../../libs/priors /opt/aframe/libs/priors
-../../libs/cosmology /opt/aframe/libs/cosmology
 ../../hermes /opt/aframe/hermes
 ../../aframe /opt/aframe/aframe
 ../../pyproject.toml /opt/aframe/pyproject.toml
