+++ projects/plots/plots/legacy/gwtc3.py
@@ -8,8 +8,9 @@ import h5py
 import numpy as np
 import scipy.stats as stats
 from astropy.utils.data import download_file
-from cosmology.cosmology import DEFAULT_COSMOLOGY
 from tqdm import tqdm
+
+from utils.cosmology import DEFAULT_COSMOLOGY
 
 catalog_results = {
     "GstLAL": {
