+++ projects/plots/plots/legacy/main.py
@@ -7,7 +7,6 @@ import jsonargparse
 import numpy as np
 from bokeh.io import save
 from bokeh.layouts import gridplot
-from cosmology.cosmology import DEFAULT_COSMOLOGY, get_astrophysical_volume
 
 from ledger.events import EventSet, RecoveredInjectionSet
 from ledger.injections import InjectionParameterSet
@@ -15,6 +14,7 @@ from plots.legacy import compute, tools
 from plots.legacy.gwtc3 import main as gwtc3_pipeline_sv
 from plots.vetos import VETO_CATEGORIES, VetoParser, get_catalog_vetos
 from priors.priors import log_normal_masses
+from utils.cosmology import DEFAULT_COSMOLOGY, get_astrophysical_volume
 from utils.logging import configure_logging
 
 logging.getLogger("urllib3").setLevel(logging.WARNING)
